requirements.txt
tensorflow==2.16.1
keras==3.3.3
numpy==1.26.4
matplotlib==3.7.5
pillow==9.5.0
flask
gunicorn==23.0.0
# python==3.10